auth-demo/pyproject.toml
[project]
name = "auth-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.0.4",
    "dash-iconify>=0.1.2",
    "dash-mantine-components>=2.0.0",
    "databricks-sdk>=0.56.0",
    "databricks-sql-connector>=4.0.3",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]
